Changes buffer size type to size_t

+++ dlib/image_loader/jpeg_loader.cpp
@@ -49,7 +49,7 @@ namespace dlib
 // ----------------------------------------------------------------------------------------
     
     jpeg_loader::
-    jpeg_loader( unsigned char* imgbuffer, long imgbuffersize ) : height_( 0 ), width_( 0 ), output_components_(0)
+    jpeg_loader( unsigned char* imgbuffer, size_t imgbuffersize ) : height_( 0 ), width_( 0 ), output_components_(0)
     {
         read_image( NULL, imgbuffer, imgbuffersize );
     }
@@ -109,7 +109,7 @@ namespace dlib
 
 // ----------------------------------------------------------------------------------------
 
-    void jpeg_loader::read_image( FILE * file, unsigned char* imgbuffer, long imgbuffersize )
+    void jpeg_loader::read_image( FILE * file, unsigned char* imgbuffer, size_t imgbuffersize )
     {
         
         jpeg_decompress_struct cinfo;
